Scroll thumb code performance improvements
